+++ src/services/userService.ts
@@ -2,14 +2,27 @@ import axios from 'axios';
 import userRepository from '../repositories/userRepository';
 import { UserDTO } from '../schemas/user/dto/userDTO';
 import { IUser } from '../schemas/user/IUser';
-import { NotFoundError } from '../utils/api-errors';
+import { BadRequestError, NotFoundError } from '../utils/api-errors';
+import { isObjectIdOrHexString } from 'mongoose';
 
 class UserService {
-  public async findAll(): Promise<IUser[]> {
-    return userRepository.findAll();
+  public async findAll(query: any): Promise<IUser[]> {
+    const { page = 1, limit = 100 } = query;
+    const queryObj = { ...query };
+
+    const excludedFields = ['select', 'sort', 'page', 'limit', 'fields'];
+    excludedFields.forEach((el) => delete queryObj[el]);
+
+    let queryStr = JSON.stringify(queryObj);
+    queryStr = queryStr.replace(/\b(gte|gt|lte|lt)\b/g, (match) => `$${match}`);
+
+    return userRepository.findAll(queryStr, page, limit);
   }
 
   public async findById(id: string): Promise<IUser | null> {
+    if (!isObjectIdOrHexString(id)) {
+      throw new BadRequestError('Invalid id provided');
+    }
     const user = await userRepository.findById(id);
 
     if (!user) {
@@ -36,6 +49,10 @@ class UserService {
   }
 
   public async delete(id: string): Promise<any> {
+    if (!isObjectIdOrHexString(id)) {
+      throw new BadRequestError('Invalid id provided');
+    }
+
     const doc = await userRepository.delete(id);
 
     if (!doc) {
@@ -44,6 +61,10 @@ class UserService {
   }
 
   public async update(id: string, userDto: UserDTO): Promise<IUser | null> {
+    if (!isObjectIdOrHexString(id)) {
+      throw new BadRequestError('Invalid id provided');
+    }
+
     const updatedUser = await userRepository.update(id, userDto);
 
     if (!updatedUser) {
